+++ src/components/Sidebar.vue
@@ -273,6 +273,7 @@ nav {
   pointer-events: none;
   transition: all 0.4s ease;
   background: rgba(0, 0, 0, 0.3);
+  z-index: 50;
 }
 
 nav .sidebar {